session code out, index.py edit, main.py cors edit
