+++ app/routers/index.py
@@ -1,106 +1,16 @@
-from fastapi import HTTPException,APIRouter,Request,Response,Depends
+from fastapi import APIRouter,Response,Depends
 from fastapi.responses import JSONResponse
 
 #const
 from common.consts import IMAPADDRESS
 from models.mailAnalysis import UserIn,responAna
 
-import numpy as np
-
 #email
-import pandas as pd
 import imaplib
-from enum import Enum
-import email
-from email import policy
-
-#response model, response body
-from pydantic import BaseModel, EmailStr
-from typing import Union,List,Optional
-
-
-#Session Backend
-from uuid import UUID,uuid4
-from fastapi_sessions.backends.implementations import InMemoryBackend
-from fastapi_sessions.session_verifier import SessionVerifier
-from fastapi_sessions.frontends.implementations import SessionCookie, CookieParameters
-
 
 router = APIRouter()
 
-#session - 47 ~ 117 lines
-class SessionData(BaseModel):
-    userinfo :UserIn
-    platform : str
-    
-cookie_params = CookieParameters()
-
-# Uses UUID
-cookie = SessionCookie(
-    cookie_name="cookie",
-    identifier="general_verifier",
-    auto_error=True,
-    secret_key="DONOTUSE",
-    cookie_params=cookie_params,
-)
-
-backend= InMemoryBackend[UUID,SessionData]()
-
-class BasicVerifier(SessionVerifier[UUID, SessionData]):
-    def __init__(
-        self,
-        *,
-        identifier: str,
-        auto_error: bool,
-        backend: InMemoryBackend[UUID, SessionData],
-        auth_http_exception: HTTPException,
-    ):
-        self._identifier = identifier
-        self._auto_error = auto_error
-        self._backend = backend
-        self._auth_http_exception = auth_http_exception
-
-    @property
-    def identifier(self):
-        return self._identifier
-
-    @property
-    def backend(self):
-        return self._backend
-
-    @property
-    def auto_error(self):
-        return self._auto_error
-
-    @property
-    def auth_http_exception(self):
-        return self._auth_http_exception
-
-    def verify_session(self, model: SessionData) -> bool:
-        """If the session exists, it is valid"""
-        return True
-
-
-verifier = BasicVerifier(
-    identifier="general_verifier",
-    auto_error=True,
-    backend=backend,
-    auth_http_exception=HTTPException(status_code=403, detail="invalid session"),
-)
-
-@router.get("/whoami", dependencies=[Depends(cookie)])
-async def whoami(session_data: SessionData = Depends(verifier)):
-    #session_data.platform
-    return session_data.platform
-
-@router.get("/create")
-async def create_session(response: Response):
-    session = uuid4()
-    data = SessionData(platform="hello")
-    await backend.create(session, data)
-    cookie.attach_to_response(response, session)
-    return "created session for"
-
+#서버 연결 Test용
 @router.get("/")
 async def read_test():
     return {"Hello":"World"}
@@ -120,12 +30,10 @@ async def access_mail(item: UserIn,response:Response):
     except imaplib.IMAP4.error as e:
         return JSONResponse(status_code=404, content={"message":"User ID or Password is invalid"})
 
-   #로그인 성공 -> Session에 정보 저장
-
-    #request.session["id"]=item.inputId     
-    #print(request.session["id"])
+   #로그인 성공, 분석 시작
 
     msg={"data":{
+            #나에게 메일을 가장 많이 보내는 사람 Top5
             "sender": [
                     {"name": "링커리어","count":600},
                     {"name": "하나투어","count":400},
@@ -133,6 +41,7 @@ async def access_mail(item: UserIn,response:Response):
                     {"name": "NEWNEEK","count":250},
                     {"name": "순살브리핑","count":100}]
             ,
+            # 메일을 받는 비율 
             "ratio":  [
                     {"name": "광고","count":100},
                     {"name": "구독","count":600},
@@ -140,6 +49,7 @@ async def access_mail(item: UserIn,response:Response):
                     {"name": "B","count":250},
                     {"name": "C","count":100}]
             ,
+            # WordCloud용 단어 최소 15개. 영어 5개, 한국어 10개
             "topic":  [
                     {"name": "취업","count":500},
                     {"name": "개발","count":300},
@@ -147,6 +57,7 @@ async def access_mail(item: UserIn,response:Response):
                     {"name": "광운대학교","count":119},
                     {"name": "유튜브","count":50}]
             ,
+            #삭제 권장 메일 ( 여러개 갈 수도 있음.)
             "delete":["응암정보도서관<ealibsend@ealib.or.kr>","UPPITY<moneyletter@uppity.co.kr>",
             "Trip.com<kr_hotel@trip.com>"]
         }}
